comment_analysis/requirements.txt
requests
langchain 
langchain-google-genai 
python-dotenv
transformers
torch
streamlit
python-dotenv
matplotlib
plotly
wordcloud 